+++ fon/FormantGrid_def.h
@@ -1,6 +1,6 @@
 /* FormantGrid_def.h
  *
- * Copyright (C) 2008-2011,2015 Paul Boersma & David Weenink
+ * Copyright (C) 2008-2011,2015,2017 Paul Boersma & David Weenink
  *
  * This code is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -25,13 +25,13 @@ oo_DEFINE_CLASS (FormantGrid, Function)
 	#if oo_DECLARING
 		bool v_hasGetVector ()
 			override { return true; }
-		double v_getVector (long irow, long icol)
+		double v_getVector (integer irow, integer icol)
 			override;
 		bool v_hasGetFunction1 ()
 			override { return true; }
-		double v_getFunction1 (long irow, double x)
+		double v_getFunction1 (integer irow, double x)
 			override;
-		const char32 * v_getUnitText (long ilevel, int /* unit */, unsigned long /* flags */)
+		const char32 * v_getUnitText (integer ilevel, int /* unit */, uint32 /* flags */)
 			override { return ilevel & 1 ? U"Formant (Hz)" : U"Bandwidth (Hz)"; }
 		void v_shiftX (double xfrom, double xto)
 			override;
