+++ stat/Table_def.h
@@ -77,15 +77,15 @@ oo_DEFINE_CLASS (Table, Daata)
 			override { return numberOfColumns; }
 		bool v_hasGetColStr ()
 			override { return true; }
-		const char32 * v_getColStr (long columnNumber)
+		const char32 * v_getColStr (integer columnNumber)
 			override;
 		bool v_hasGetMatrix ()
 			override { return true; }
-		double v_getMatrix (long rowNumber, long columnNumber)
+		double v_getMatrix (integer rowNumber, integer columnNumber)
 			override;
 		bool v_hasGetMatrixStr ()
 			override { return true; }
-		const char32 * v_getMatrixStr (long rowNumber, long columnNumber)
+		const char32 * v_getMatrixStr (integer rowNumber, integer columnNumber)
 			override;
 		bool v_hasGetColIndex ()
 			override { return true; }
