+++ sys/oo_WRITE_BINARY.h
@@ -64,7 +64,7 @@
 		binput##storage (our x [i], f);
 
 #define oo_STRINGx_VECTOR(storage,x,min,max)  \
-	for (long i = min; i <= max; i ++) \
+	for (integer i = min; i <= max; i ++) \
 		binput##storage (our x [i], f);
 
 #define oo_STRUCT(Type,x)  \
@@ -79,12 +79,12 @@
 		our x [i]. writeBinary (f);
 
 #define oo_STRUCT_VECTOR_FROM(Type,x,min,max)  \
-	for (long i = min; i <= max; i ++) \
+	for (integer i = min; i <= max; i ++) \
 		our x [i]. writeBinary (f);
 
 #define oo_STRUCT_MATRIX_FROM(Type,x,row1,row2,col1,col2)  \
-	for (long i = row1; i <= row2; i ++) \
-		for (long j = col1; j <= col2; j ++) \
+	for (integer i = row1; i <= row2; i ++) \
+		for (integer j = col1; j <= col2; j ++) \
 			our x [i] [j]. writeBinary (f);
 
 #define oo_AUTO_OBJECT(Class,version,x)  \
@@ -94,7 +94,7 @@
 
 #define oo_COLLECTION_OF(Class,x,ItemClass,version)  \
 	binputi32 (our x.size, f); \
-	for (long i = 1; i <= our x.size; i ++) { \
+	for (integer i = 1; i <= our x.size; i ++) { \
 		ItemClass data = our x.at [i]; \
 		data -> struct##ItemClass :: v_writeBinary (f); \
 	}
@@ -102,7 +102,7 @@
 #define oo_AUTO_COLLECTION(Class,x,ItemClass,version)  \
 	binputi32 (our x ? our x->size : 0, f); \
 	if (our x) { \
-		for (long i = 1; i <= our x->size; i ++) { \
+		for (integer i = 1; i <= our x->size; i ++) { \
 			ItemClass data = our x->at [i]; \
 			data -> struct##ItemClass :: v_writeBinary (f); \
 		} \
